Fix CSApprox when run inside xterm or screen/tmux
Currently on Ubuntu and probably other distros, 256 colors is not enabled
in terminal Vim. Because of this CSApprox issues the following warning
every time the user starts vim from the within xterm or a screen/tmux
session:

> CSApprox skipped; terminal only has 8 colors, not 88/256
> Try checking :help csapprox-terminal for workarounds

This fix adds a check of the terminal being used and if it's xterm or
screen (tmux also identifies itself as such) then we set t_Co to
256 colors. I think thinks is reasonable as xterm has supported 256
colors for about 13 years now.