add backup .vim files in xp-installer
I believe the original bootstrap.sh checks if the files to be backed up
exist and are not symbolic links. In case of XP I think that not backing
up if "..\.spf13-vim-3" exists would probably be sufficient to save
updating users the wait.
If they decided to use the distribution the .vimrc should not be edited
manually anyway.